requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymemcache==3.4.1
pyTelegramBotAPI==3.7.7
python-dotenv==0.16.0
python-gitlab==2.6.0
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
